+++ opentelemetry-sdk/src/logs/error.rs
@@ -1,6 +1,9 @@
+// Re-export ShutdownError
+pub use crate::error::ShutdownError;
+
 use crate::export::ExportError;
 
-use std::{sync::PoisonError, time::Duration};
+use std::time::Duration;
 use thiserror::Error;
 
 /// Describe the result of operations in log SDK.
@@ -18,13 +21,15 @@ pub enum LogError {
     #[error("Exporter timed out after {} seconds", .0.as_secs())]
     ExportTimedOut(Duration),
 
+    /// The export client failed while holding the client lock. It is not
+    /// possible to complete the shutdown and a retry will not help.
+    /// This is something that should not happen and should likely emit some diagnostic.
+    #[error("export client failed while holding lock; cannot retry.")]
+    ClientFailed(String),
+
     /// Processor is already shutdown
     #[error("{0} already shutdown")]
     AlreadyShutdown(String),
-
-    /// Mutex lock poisoning
-    #[error("mutex lock poisioning for {0}")]
-    MutexPoisoned(String),
 
     /// Other errors propagated from log SDK that weren't covered above.
     #[error(transparent)]
@@ -52,11 +57,6 @@ impl From<&'static str> for LogError {
     }
 }
 
-impl<T> From<PoisonError<T>> for LogError {
-    fn from(err: PoisonError<T>) -> Self {
-        LogError::Other(err.to_string().into())
-    }
-}
 /// Wrap type for string
 #[derive(Error, Debug)]
 #[error("{0}")]