+++ opentelemetry-stdout/src/logs/exporter.rs
@@ -1,6 +1,6 @@
 use chrono::{DateTime, Utc};
 use core::fmt;
-use opentelemetry_sdk::export::logs::LogBatch;
+use opentelemetry_sdk::export::logs::{LogBatch, ShutdownResult};
 use opentelemetry_sdk::logs::LogResult;
 use opentelemetry_sdk::Resource;
 use std::sync::atomic;
@@ -63,8 +63,9 @@ impl opentelemetry_sdk::export::logs::LogExporter for LogExporter {
         }
     }
 
-    fn shutdown(&mut self) {
+    fn shutdown(&mut self) -> ShutdownResult {
         self.is_shutdown.store(true, atomic::Ordering::SeqCst);
+        Ok(())
     }
 
     fn set_resource(&mut self, res: &opentelemetry_sdk::Resource) {